refactor listener.py into inference.py
